requirements.txt
tensorflow
opencv-python
numpy
pillow
matplotlib
tkinter
